Safari-609.1.11.1/LayoutTests/media/video-play-pause-events-expected.txt


Test that calling play() and pause() triggers async play, no timeupdate event and pause event.

RUN(handlePromise(video.play()))
RUN(video.pause())
SCRIPT DONE
EVENT(play)
EVENT(waiting)
EVENT(pause)
EXPECTED (video.paused == 'true') OK
END OF TEST
